[te] Benchmark vml-based logit
Pull Request resolved: https://github.com/pytorch/pytorch/pull/51771

This benchmarks an NNC implementation of logit based on VML's log
implementation.

It's a modest improvement over the sleef algorithm, but seems to be a bit
slower than aten (at larger sizes), and I'm not totally sure why, since you'd
think a fused logit kernel would be better than doing clamp/sub/div, followed
by log.  And yet...

Note that it's important to vectorize this kernel by 16, even on an 8-wide AVX2
machine; I suspect that it's needed to give the scheduler enough freedom to
fill up both FMA pipes to avoid stalling on fpdiv or (maybe) memory.
ghstack-source-id: 121366312

Differential Revision: [D26272325](https://our.internmc.facebook.com/intern/diff/D26272325/)